+++ packages/gas-fee-controller/src/GasFeeController.ts
@@ -25,13 +25,13 @@ import { v1 as random } from 'uuid';
 import determineGasFeeCalculations from './determineGasFeeCalculations';
 import fetchGasEstimatesViaEthFeeHistory from './fetchGasEstimatesViaEthFeeHistory';
 import {
+  calculateTimeEstimate,
   fetchGasEstimates,
   fetchLegacyGasPriceEstimates,
   fetchEthGasPriceEstimate,
-  calculateTimeEstimate,
 } from './gas-util';
 
-export const LEGACY_GAS_PRICES_API_URL = `https://api.metaswap.codefi.network/gasPrices`;
+export const GAS_API_BASE_URL = 'https://gas.api.infura.io';
 
 export type unknownString = 'unknown';
 
@@ -274,6 +274,8 @@ export class GasFeeController extends StaticIntervalPollingController<
 
   private readonly getCurrentAccountEIP1559Compatibility;
 
+  private readonly infuraAPIKey: string;
+
   private currentChainId;
 
   private ethQuery?: EthQuery;
@@ -299,11 +301,9 @@ export class GasFeeController extends StaticIntervalPollingController<
    * @param options.getProvider - Returns a network provider for the current network.
    * @param options.onNetworkDidChange - A function for registering an event handler for the
    * network state change event.
-   * @param options.legacyAPIEndpoint - The legacy gas price API URL. This option is primarily for
-   * testing purposes.
-   * @param options.EIP1559APIEndpoint - The EIP-1559 gas price API URL.
    * @param options.clientId - The client ID used to identify to the gas estimation API who is
    * asking for estimates.
+   * @param options.infuraAPIKey - The Infura API key used for infura API requests.
    */
   constructor({
     interval = 15000,
@@ -315,9 +315,8 @@ export class GasFeeController extends StaticIntervalPollingController<
     getCurrentNetworkLegacyGasAPICompatibility,
     getProvider,
     onNetworkDidChange,
-    legacyAPIEndpoint = LEGACY_GAS_PRICES_API_URL,
-    EIP1559APIEndpoint,
     clientId,
+    infuraAPIKey,
   }: {
     interval?: number;
     messenger: GasFeeMessenger;
@@ -328,9 +327,8 @@ export class GasFeeController extends StaticIntervalPollingController<
     getChainId?: () => Hex;
     getProvider: () => ProviderProxy;
     onNetworkDidChange?: (listener: (state: NetworkState) => void) => void;
-    legacyAPIEndpoint?: string;
-    EIP1559APIEndpoint: string;
     clientId?: string;
+    infuraAPIKey: string;
   }) {
     super({
       name,
@@ -348,9 +346,10 @@ export class GasFeeController extends StaticIntervalPollingController<
     this.getCurrentAccountEIP1559Compatibility =
       getCurrentAccountEIP1559Compatibility;
     this.#getProvider = getProvider;
-    this.EIP1559APIEndpoint = EIP1559APIEndpoint;
-    this.legacyAPIEndpoint = legacyAPIEndpoint;
+    this.EIP1559APIEndpoint = `${GAS_API_BASE_URL}/networks/<chain_id>/suggestedGasFees`;
+    this.legacyAPIEndpoint = `${GAS_API_BASE_URL}/networks/<chain_id>/gasPrices`;
     this.clientId = clientId;
+    this.infuraAPIKey = infuraAPIKey;
 
     this.ethQuery = new EthQuery(this.#getProvider());
 
@@ -473,6 +472,7 @@ export class GasFeeController extends StaticIntervalPollingController<
       calculateTimeEstimate,
       clientId: this.clientId,
       ethQuery,
+      infuraAPIKey: this.infuraAPIKey,
     });
 
     if (shouldUpdateState) {