+++ packages/gas-fee-controller/src/determineGasFeeCalculations.test.ts
@@ -40,6 +40,8 @@ const mockedFetchGasEstimatesViaEthFeeHistory =
     Parameters<typeof fetchGasEstimatesViaEthFeeHistory>
   >;
 
+const INFURA_API_KEY_MOCK = 'test';
+
 /**
  * Builds mock data for the `fetchGasEstimates` function. All of the data here is filled in to make
  * the gas fee estimation code function in a way that represents a reasonably happy path; it does
@@ -132,6 +134,7 @@ describe('determineGasFeeCalculations', () => {
     calculateTimeEstimate: mockedCalculateTimeEstimate,
     clientId: 'some-client-id',
     ethQuery: {},
+    infuraAPIKey: INFURA_API_KEY_MOCK,
   };
 
   describe('when isEIP1559Compatible is true', () => {