+++ packages/gas-fee-controller/src/determineGasFeeCalculations.ts
@@ -31,6 +31,7 @@ import { GAS_ESTIMATE_TYPES } from './GasFeeController';
  * @param args.calculateTimeEstimate - A function that determine time estimate bounds.
  * @param args.clientId - An identifier that an API can use to know who is asking for estimates.
  * @param args.ethQuery - An EthQuery instance we can use to talk to Ethereum directly.
+ * @param args.infuraAPIKey - Infura API key to use for requests to Infura.
  * @returns The gas fee calculations.
  */
 export default async function determineGasFeeCalculations({
@@ -45,11 +46,13 @@ export default async function determineGasFeeCalculations({
   calculateTimeEstimate,
   clientId,
   ethQuery,
+  infuraAPIKey,
 }: {
   isEIP1559Compatible: boolean;
   isLegacyGasAPICompatible: boolean;
   fetchGasEstimates: (
     url: string,
+    infuraAPIKey: string,
     clientId?: string,
   ) => Promise<GasFeeEstimates>;
   fetchGasEstimatesUrl: string;
@@ -60,6 +63,7 @@ export default async function determineGasFeeCalculations({
   ) => Promise<GasFeeEstimates>;
   fetchLegacyGasPriceEstimates: (
     url: string,
+    infuraAPIKey: string,
     clientId?: string,
   ) => Promise<LegacyGasPriceEstimate>;
   fetchLegacyGasPriceEstimatesUrl: string;
@@ -75,12 +79,17 @@ export default async function determineGasFeeCalculations({
   // TODO: Replace `any` with type
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ethQuery: any;
+  infuraAPIKey: string;
 }): Promise<GasFeeCalculations> {
   try {
     if (isEIP1559Compatible) {
       let estimates: GasFeeEstimates;
       try {
-        estimates = await fetchGasEstimates(fetchGasEstimatesUrl, clientId);
+        estimates = await fetchGasEstimates(
+          fetchGasEstimatesUrl,
+          infuraAPIKey,
+          clientId,
+        );
       } catch {
         estimates = await fetchGasEstimatesViaEthFeeHistory(ethQuery);
       }
@@ -99,6 +108,7 @@ export default async function determineGasFeeCalculations({
     } else if (isLegacyGasAPICompatible) {
       const estimates = await fetchLegacyGasPriceEstimates(
         fetchLegacyGasPriceEstimatesUrl,
+        infuraAPIKey,
         clientId,
       );
       return {