+++ packages/transaction-controller/src/helpers/GasFeePoller.ts
@@ -1,6 +1,6 @@
-import type EthQuery from '@metamask/eth-query';
+import EthQuery from '@metamask/eth-query';
 import type { GasFeeState } from '@metamask/gas-fee-controller';
-import type { NetworkClientId } from '@metamask/network-controller';
+import type { NetworkClientId, Provider } from '@metamask/network-controller';
 import type { Hex } from '@metamask/utils';
 import { createModuleLogger } from '@metamask/utils';
 import EventEmitter from 'events';
@@ -28,9 +28,9 @@ export class GasFeePoller {
 
   #gasFeeFlows: GasFeeFlow[];
 
-  #getEthQuery: (chainId: Hex, networkClientId?: NetworkClientId) => EthQuery;
-
   #getGasFeeControllerEstimates: () => Promise<GasFeeState>;
+
+  #getProvider: (chainId: Hex, networkClientId?: NetworkClientId) => Provider;
 
   #getTransactions: () => TransactionMeta[];
 
@@ -44,31 +44,31 @@ export class GasFeePoller {
    * Constructs a new instance of the GasFeePoller.
    * @param options - The options for this instance.
    * @param options.gasFeeFlows - The gas fee flows to use to obtain suitable gas fees.
-   * @param options.getEthQuery - Callback to obtain an EthQuery instance.
    * @param options.getGasFeeControllerEstimates - Callback to obtain the default fee estimates.
+   * @param options.getProvider - Callback to obtain a provider instance.
    * @param options.getTransactions - Callback to obtain the transaction data.
    * @param options.layer1GasFeeFlows - The layer 1 gas fee flows to use to obtain suitable layer 1 gas fees.
    * @param options.onStateChange - Callback to register a listener for controller state changes.
    */
   constructor({
     gasFeeFlows,
-    getEthQuery,
     getGasFeeControllerEstimates,
+    getProvider,
     getTransactions,
     layer1GasFeeFlows,
     onStateChange,
   }: {
     gasFeeFlows: GasFeeFlow[];
-    getEthQuery: (chainId: Hex, networkClientId?: NetworkClientId) => EthQuery;
     getGasFeeControllerEstimates: () => Promise<GasFeeState>;
+    getProvider: (chainId: Hex, networkClientId?: NetworkClientId) => Provider;
     getTransactions: () => TransactionMeta[];
     layer1GasFeeFlows: Layer1GasFeeFlow[];
     onStateChange: (listener: () => void) => void;
   }) {
     this.#gasFeeFlows = gasFeeFlows;
     this.#layer1GasFeeFlows = layer1GasFeeFlows;
-    this.#getEthQuery = getEthQuery;
     this.#getGasFeeControllerEstimates = getGasFeeControllerEstimates;
+    this.#getProvider = getProvider;
     this.#getTransactions = getTransactions;
 
     onStateChange(() => {
@@ -134,7 +134,7 @@ export class GasFeePoller {
   async #updateTransactionSuggestedFees(transactionMeta: TransactionMeta) {
     const { chainId, networkClientId } = transactionMeta;
 
-    const ethQuery = this.#getEthQuery(chainId, networkClientId);
+    const ethQuery = new EthQuery(this.#getProvider(chainId, networkClientId));
     const gasFeeFlow = getGasFeeFlow(transactionMeta, this.#gasFeeFlows);
 
     if (!gasFeeFlow) {
@@ -184,10 +184,10 @@ export class GasFeePoller {
 
   async #updateTransactionLayer1GasFee(transactionMeta: TransactionMeta) {
     const { chainId, networkClientId } = transactionMeta;
-    const ethQuery = this.#getEthQuery(chainId, networkClientId);
+    const provider = this.#getProvider(chainId, networkClientId);
 
     await updateTransactionLayer1GasFee({
-      ethQuery,
+      provider,
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
       transactionMeta,
     });