+++ packages/transaction-controller/src/utils/layer1-gas-fee-flow.test.ts
@@ -1,5 +1,4 @@
-import { query } from '@metamask/controller-utils';
-import type EthQuery from '@metamask/eth-query';
+import type { Provider } from '@metamask/network-controller';
 import type { Hex } from '@metamask/utils';
 
 import {
@@ -38,8 +37,8 @@ function createLayer1GasFeeFlowMock({
 }
 
 describe('updateTransactionLayer1GasFee', () => {
-  const queryMock = query as unknown as EthQuery;
   let layer1GasFeeFlowsMock: jest.Mocked<Layer1GasFeeFlow[]>;
+  let providerMock: Provider;
   let transactionMetaMock: TransactionMeta;
 
   beforeEach(() => {
@@ -53,6 +52,7 @@ describe('updateTransactionLayer1GasFee', () => {
         layer1Fee: LAYER1_GAS_FEE_VALUE_MATCH_MOCK,
       }),
     ];
+    providerMock = {} as Provider;
     transactionMetaMock = {
       id: '1',
       chainId: '0x123',
@@ -66,8 +66,8 @@ describe('updateTransactionLayer1GasFee', () => {
 
   it('updates given transaction layer1GasFee property', async () => {
     await updateTransactionLayer1GasFee({
-      ethQuery: queryMock,
       layer1GasFeeFlows: layer1GasFeeFlowsMock,
+      provider: providerMock,
       transactionMeta: transactionMetaMock,
     });
 
@@ -77,7 +77,7 @@ describe('updateTransactionLayer1GasFee', () => {
     expect(unmatchingLayer1GasFeeFlow.getLayer1Fee).not.toHaveBeenCalled();
 
     expect(matchingLayer1GasFeeFlow.getLayer1Fee).toHaveBeenCalledWith({
-      ethQuery: queryMock,
+      provider: providerMock,
       transactionMeta: transactionMetaMock,
     });
 
@@ -96,13 +96,13 @@ describe('updateTransactionLayer1GasFee', () => {
       );
 
       await updateTransactionLayer1GasFee({
-        ethQuery: queryMock,
-        transactionMeta: transactionMetaMock,
         layer1GasFeeFlows: layer1GasFeeFlowsMock,
+        provider: providerMock,
+        transactionMeta: transactionMetaMock,
       });
 
       expect(matchingLayer1GasFeeFlow.getLayer1Fee).toHaveBeenCalledWith({
-        ethQuery: queryMock,
+        provider: providerMock,
         transactionMeta: transactionMetaMock,
       });
       expect(transactionMetaMock.layer1GasFee).toBeUndefined();
@@ -116,9 +116,9 @@ describe('updateTransactionLayer1GasFee', () => {
       layer1GasFeeFlowsMock = [unmatchingLayer1GasFeeFlow];
 
       await updateTransactionLayer1GasFee({
-        ethQuery: queryMock,
-        transactionMeta: transactionMetaMock,
         layer1GasFeeFlows: layer1GasFeeFlowsMock,
+        provider: providerMock,
+        transactionMeta: transactionMetaMock,
       });
 
       expect(unmatchingLayer1GasFeeFlow.getLayer1Fee).not.toHaveBeenCalled();